refactor(cli): move import and export under session
