+++ packages/cli/src/commands/handlers/session/export.ts
@@ -3,14 +3,14 @@ import { OpenCode } from "@opencode/client"
 import { Service } from "@opencode/client/effect/service"
 import { Effect, Option } from "effect"
 import { EOL } from "node:os"
-import { Commands } from "../commands"
-import { Runtime } from "../../framework/runtime"
-import { ServerConnection } from "../../services/server-connection"
-import { errorMessage } from "../../util/error"
+import { Commands } from "../../commands"
+import { Runtime } from "../../../framework/runtime"
+import { ServerConnection } from "../../../services/server-connection"
+import { errorMessage } from "../../../util/error"
 
 export default Runtime.handler(
-  Commands.commands.export,
-  Effect.fn("cli.export")((input) =>
+  Commands.commands.session.commands.export,
+  Effect.fn("cli.session.export")((input) =>
     Effect.gen(function* () {
       const requested = Option.getOrUndefined(input.session)
       if (!requested && !process.stdin.isTTY) {
