+++ packages/cli/test/import-export.test.ts
@@ -71,13 +71,14 @@ test("export is raw by default and supports explicit sanitization", async () => 
   })
 
   try {
-    const [stdout, , exitCode] = await run(["export", info.id, "--server", server.url.toString()])
+    const [stdout, , exitCode] = await run(["session", "export", info.id, "--server", server.url.toString()])
     const exported = JSON.parse(stdout)
 
     expect(exitCode).toBe(0)
     expect(exported).toEqual(transfer)
 
     const [sanitized, , sanitizedExitCode] = await run([
+      "session",
       "export",
       info.id,
       "--sanitize",
@@ -110,7 +111,7 @@ test("export requires a session outside an interactive terminal", async () => {
   })
 
   try {
-    const [stdout, stderr, exitCode] = await run(["export", "--server", server.url.toString()])
+    const [stdout, stderr, exitCode] = await run(["session", "export", "--server", server.url.toString()])
 
     expect(exitCode).toBe(1)
     expect(stdout).toBe("")
@@ -138,7 +139,7 @@ test("export reports a missing session without a stack trace", async () => {
   })
 
   try {
-    const [stdout, stderr, exitCode] = await run(["export", sessionID, "--server", server.url.toString()])
+    const [stdout, stderr, exitCode] = await run(["session", "export", sessionID, "--server", server.url.toString()])
 
     expect(exitCode).toBe(1)
     expect(stdout).toBe("")
@@ -173,7 +174,15 @@ test("import validates a file and sends it to the resolved location", async () =
   })
 
   try {
-    const [stdout, , exitCode] = await run(["import", file, "--directory", root, "--server", server.url.toString()])
+    const [stdout, , exitCode] = await run([
+      "session",
+      "import",
+      file,
+      "--directory",
+      root,
+      "--server",
+      server.url.toString(),
+    ])
 
     expect(exitCode).toBe(0)
     expect(stdout).toBe(`Imported session: ${info.id}${os.EOL}`)
@@ -205,7 +214,7 @@ test("import reports an existing session without a stack trace", async () => {
   })
 
   try {
-    const [stdout, stderr, exitCode] = await run(["import", file, "--server", server.url.toString()])
+    const [stdout, stderr, exitCode] = await run(["session", "import", file, "--server", server.url.toString()])
 
     expect(exitCode).toBe(0)
     expect(stdout).toBe("")
